fieldmanager: Move structured benchmarks to their own file
